test(tmux): gate TestSession_SetAndGetEnvironment behind hostsensitive tag (#969)
The test depends on a live external tmux server and is sensitive to
tmux's per-session environment table being clean of prior state — observed
flaky on hosts with lingering tmux servers. Move it behind the
hostsensitive build tag so default pre-push and CI runs stay green.

Refs: #969

Committed by Ashesh Goplani
