+++ internal/tmux/tmux_test.go
@@ -2264,35 +2264,7 @@ func TestSessionLogFile(t *testing.T) {
 	assert.True(t, strings.HasSuffix(logFile, ".log"))
 }
 
-func TestSession_SetAndGetEnvironment(t *testing.T) {
-	skipIfNoTmuxServer(t)
-
-	// Create a test session
-	sess := NewSession("env-test", "/tmp")
-
-	// Start the session (required for environment to work)
-	err := sess.Start("")
-	if err != nil {
-		t.Fatalf("Failed to start session: %v", err)
-	}
-	defer func() { _ = sess.Kill() }()
-
-	// Test setting environment
-	err = sess.SetEnvironment("TEST_VAR", "test_value_123")
-	if err != nil {
-		t.Fatalf("SetEnvironment failed: %v", err)
-	}
-
-	// Test getting environment
-	value, err := sess.GetEnvironment("TEST_VAR")
-	if err != nil {
-		t.Fatalf("GetEnvironment failed: %v", err)
-	}
-
-	if value != "test_value_123" {
-		t.Errorf("GetEnvironment = %q, want %q", value, "test_value_123")
-	}
-}
+// TestSession_SetAndGetEnvironment moved to tmux_hostsensitive_test.go (#969).
 
 func TestSession_GetEnvironment_NotFound(t *testing.T) {
 	skipIfNoTmuxServer(t)
